# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ario Web Services",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1"
        },
        {
            "url": "https://ario.jcoders.ir/api/v1"
        }
    ],
    "paths": {
...
        "/admin/orders": {
            "get": {
                "tags": [
                    "سفارشات - پنل مدیریت"
                ],
                "summary": "لیست سفارشات بصورت صفحه بندی",
                "description": "ثبت سفارش",
                "operationId": "5478286f8e1b7aa4fe445da47494d20f",
                "parameters": [
                    {
                        "name": "order",
                        "in": "query",
                        "description": "شماره سفارش",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "user_name",
                        "in": "query",
                        "description": "نام سفارش دهنده",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "شناسه سفارش دهنده",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": " 1401/8/24 - از تاریخ",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": " 1401/8/24 - تا تاریخ",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "وضعیت",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Pending",
                                "Success",
                                "Canceled",
                                "Fail"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "عملیات موفق",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/admin/orders/{id}": {
            "get": {
                "tags": [
                    "سفارشات - پنل مدیریت"
                ],
                "summary": "نمایش جزئیات سفارش",
                "description": "نمایش جزئیات سفارش",
                "operationId": "01fe47b2a842ba3c0ebdaee982b3c92b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "شناسه سفارش",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "عملیات موفق",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "توکن را به همراه Bearer وارد کنید.",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}